requirements.txt
Django==2.1.4
django-registration-redux==2.6
kombu==4.5.0
mysqlclient==1.4.1
netaddr==0.7.19
pyeapi==0.8.2
pytz==2018.9
django-crispy-forms==1.7.2
Pillow==5.4.1
